[react-feedback-use-case] added callback initializer to history change hook.
This hook should placed in a component that able to detect all history changes. In this app, header can detect all history changes because part of all pages.
If your app doesn't have a component like header which rendering on all pages, you should call this hook on App.js.